renamed contants to config
